Moved random bash commands for installing to install script

+++ code/install.sh
@@ -1,14 +1,23 @@
 #!/bin/bash 
 
-echo 'Installing requirements..'
+echo "Installing requirements.."
 
+pip3 install --upgrade setuptools
 pip3 install -r ../requirements.txt
 
 chmod +x panoptes.py
 
-mkdir /var/log/panoptes/
-mkdir /etc/panoptes/
+if [ ! -d /var/log/panoptes/ ]; then
+    mkdir /var/log/panoptes/
+fi
 
-echo 'Config.py created'
+if [ ! -d /etc/panoptes/ ]; then
+    mkdir /etc/panoptes/
+fi
 
-cp config.example config.py
+if [ ! -f config.py ]; then
+    echo "Config.py created"
+    cp config.example config.py
+else
+    echo "Using old config"
+fi
